hide internal i/f from public one
