+++ libs/aws/sabi-s3/src/actions/put_object/error.rs
@@ -7,9 +7,6 @@ pub enum Error {
 
     #[fail(display = "crate::core::Error > {}", 0)]
     S3CoreError(crate::core::Error),
-
-    #[fail(display = "sabi_core::Error > {}", 0)]
-    SabiCoreError(sabi_core::Error),
 }
 
 impl From<internal::Error> for Error {
@@ -21,11 +18,5 @@ impl From<internal::Error> for Error {
 impl From<crate::core::Error> for Error {
     fn from(e: crate::core::Error) -> Self {
         Error::S3CoreError(e)
-    }
-}
-
-impl From<sabi_core::Error> for Error {
-    fn from(e: sabi_core::Error) -> Self {
-        Error::SabiCoreError(e)
     }
 }
