+++ libs/aws/sabi-s3/src/core/request/mod.rs
@@ -1,0 +1,3 @@
+mod request_resource;
+pub use request_resource::RequestResource;
+pub use request_resource::ResourceLoader;
