+++ libs/aws/sabi-s3/src/internal/error.rs
@@ -4,9 +4,6 @@ pub type Result<T> = std::result::Result<T, Error>;
 
 #[derive(Fail, Debug)]
 pub enum Error {
-    #[fail(display = "FileNotFound > {}", 0)]
-    FileNotFound { path: String, description: String },
-
     #[fail(display = "region not specified")]
     RegionNotSpecified,
