+++ libs/aws/sabi-s3/src/internal/impl_async/mod.rs
@@ -7,9 +7,5 @@ pub use internal_request::InternalRequest;
 mod request_provider;
 pub use request_provider::RequestProvider;
 
-mod request_resource;
-pub use request_resource::RequestResource;
-pub use request_resource::ResourceLoader;
-
 mod error_response;
 pub use error_response::S3ErrorResponse;
